sdsdf

+++ WMS/Models/ViewAbsent.cs
@@ -52,7 +52,7 @@ namespace WMS.Models
         public Nullable<byte> TypeID { get; set; }
         public Nullable<short> CatID { get; set; }
         public Nullable<short> DeptID { get; set; }
-        public Nullable<byte> GradeID { get; set; }
+        public Nullable<short> GradeID { get; set; }
         public Nullable<short> CrewID { get; set; }
         public Nullable<short> LocID { get; set; }
         public Nullable<byte> ShiftID { get; set; }
